+++ packages/alfa-css-feature/src/media/feature/resolver.ts
@@ -15,7 +15,7 @@ export namespace Resolver {
    *
    * @remarks
    * Relative lengths in media queries are based on initial values of the
-   * associated properties. They are hard-coded here to avoid circular dependcy
+   * associated properties. They are hard-coded here to avoid circular dependency
    * to @siteimprove/alfa-style.
    */
   export function length(device: Device): Length.Resolver {
