+++ packages/alfa-css-feature/src/media/feature/width.ts
@@ -2,10 +2,9 @@ import { Length } from "@siteimprove/alfa-css";
 import { Device } from "@siteimprove/alfa-device";
 import { None, Option } from "@siteimprove/alfa-option";
 
-import { Resolver } from "../resolver";
-import { Value } from "../value";
-
 import { Media } from "./media";
+import { Resolver } from "./resolver";
+import { Value } from "./value";
 
 /**
  * {@link https://drafts.csswg.org/mediaqueries-5/#width}
