+++ packages/alfa-css-feature/tsconfig.json
@@ -9,24 +9,25 @@
     "src/condition/or.ts",
     "src/feature.ts",
     "src/index.ts",
-    "src/list.ts",
     "src/matchable.ts",
-    "src/media/comparison.ts",
-    "src/media/media.ts",
-    "src/media/height.ts",
+    "src/media/feature/comparison.ts",
+    "src/media/feature/media.ts",
+    "src/media/feature/height.ts",
+    "src/media/feature/index.ts",
+    "src/media/feature/orientation.ts",
+    "src/media/feature/resolver.ts",
+    "src/media/feature/scripting.ts",
+    "src/media/feature/value/bound.ts",
+    "src/media/feature/value/discrete.ts",
+    "src/media/feature/value/index.ts",
+    "src/media/feature/value/range.ts",
+    "src/media/feature/value/value.ts",
+    "src/media/feature/width.ts",
     "src/media/index.ts",
-    "src/media/orientation.ts",
-    "src/media/scripting.ts",
-    "src/media/width.ts",
-    "src/modifier.ts",
-    "src/query.ts",
-    "src/resolver.ts",
-    "src/type.ts",
-    "src/value/bound.ts",
-    "src/value/discrete.ts",
-    "src/value/index.ts",
-    "src/value/range.ts",
-    "src/value/value.ts",
+    "src/media/list.ts",
+    "src/media/modifier.ts",
+    "src/media/query.ts",
+    "src/media/type.ts",
     "test/media.spec.ts"
   ],
   "references": [
