+++ packages/alfa-dom/src/style/rule/media.ts
@@ -16,17 +16,17 @@ export class MediaRule extends ConditionRule<"media"> {
     return new MediaRule(condition, Array.from(rules));
   }
 
-  private readonly _queries: Feature.List;
+  private readonly _queries: Feature.Media.List;
 
   private constructor(condition: string, rules: Array<Rule>) {
     super("media", condition, rules);
 
     this._queries = Feature.parseMediaQuery(Lexer.lex(condition))
       .map(([, queries]) => queries)
-      .getOr(Feature.List.of([]));
+      .getOr(Feature.Media.List.of([]));
   }
 
-  public get queries(): Feature.List {
+  public get queries(): Feature.Media.List {
     return this._queries;
   }
